apps/gateway/requirements.txt
-r ../../requirements-shared.txt

# Gateway 特定
fastapi-limiter>=0.1.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
redis>=5.0.0
